% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBasedRecommender.R
\name{closest_geohash_vertex}
\alias{closest_geohash_vertex}
\title{Closest Geohash vertex from a graph}
\usage{
closest_geohash_vertex(graph, gh, matCoords = NULL)
}
\description{
Finds the nearest neighbor Geohash tile for a given Geo-point.
}
