% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfUSARouteSystemGraphEdges}
\alias{dfUSARouteSystemGraphEdges}
\title{USA route system graph edges}
\format{
An object of class \code{data.frame} with 136712 rows and 3 columns.
}
\source{
https://en.wikipedia.org/wiki/Interstate_Highway_System
}
\usage{
dfUSARouteSystemGraphEdges
}
\description{
A data frame with "from" and "to" columns the values of which are geohashes.
}
\keyword{datasets}
