% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBasedRecommender.R
\name{find_route_graph_paths}
\alias{find_route_graph_paths}
\title{Find route graph paths}
\usage{
find_route_graph_paths(data, graph = NULL)
}
\arguments{
\item{data}{Data frame with columns "id", "lat1", "lon1", "lat2", 'lon2".}

\item{graph}{Route system graph.
If NULL a geohashes route graph with precision 4 is used.}
}
\description{
Find shortest paths for a given data frame with start and end locations
over a given route graph.
}
\details{
Generally speaking any graph based on geohash can be used,
but the main use case of this function is to use graphs based on a route system.
}
