% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBasedRecommender.R
\name{make_route_graph}
\alias{make_route_graph}
\title{Make route graph}
\usage{
make_route_graph(precision = 5, directed = FALSE)
}
\arguments{
\item{precision}{Geohash precision to use.}

\item{directed}{Should the graph be directed or not?}
}
\description{
Makes a route graph from package's generic set of edges.
}
\details{
Package's edge data is with geohash precision 9.
Hence, `precision` should be 9 or less.
}
